+++ Example/AppClipExample/WatchApplePayExample Watch AppTests/WatchApplePayExample_Watch_AppTests.swift
@@ -1,0 +1,37 @@
+//
+//  WatchApplePayExample_Watch_AppTests.swift
+//  WatchApplePayExample Watch AppTests
+//
+//  Created by David Estes on 11/4/22.
+//  Copyright © 2022 Stripe, Inc. All rights reserved.
+//
+
+import XCTest
+@testable import WatchApplePayExample_Watch_App
+
+final class WatchApplePayExample_Watch_AppTests: XCTestCase {
+
+    override func setUpWithError() throws {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+    }
+
+    override func tearDownWithError() throws {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testExample() throws {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+        // Any test you write for XCTest can be annotated as throws and async.
+        // Mark your test throws to produce an unexpected failure when your test encounters an uncaught error.
+        // Tests marked async will run the test method on an arbitrary thread managed by the Swift runtime.
+    }
+
+    func testPerformanceExample() throws {
+        // This is an example of a performance test case.
+        self.measure {
+            // Put the code you want to measure the time of here.
+        }
+    }
+
+}