+++ StripeCore/StripeCoreTests/API Bindings/StripeCodableTest.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2021 Stripe, Inc. All rights reserved.
 //
 
+#if os(iOS)
 import Foundation
 import OHHTTPStubs
 import OHHTTPStubsSwift
@@ -210,3 +211,5 @@ class StripeAPIRequestTest: APIStubbedTestCase {
         }
     }
 }
+
+#endif