+++ StripeIdentity/StripeIdentityTests/Snapshot/VerificationFlowWebViewSnapshotTests.swift
@@ -7,6 +7,7 @@
 //
 
 @_spi(STP) import StripeCore
+@_spi(STP) import StripeUICore
 import WebKit
 import iOSSnapshotTestCase